ocr-service/requirements.txt
flask>=3.0.0
pytesseract>=0.3.10
opencv-python>=4.8.0
pdf2image>=1.17.0
pdfplumber>=0.10.0
Pillow>=10.0.0
numpy>=1.24.0
